quotron/auth-engine/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
sqlalchemy==2.0.23
pytest==7.4.3